Combine exporter and cacher phases
Combines the exporter and cacher phases into one exporter phase to reuse
reuse layer hashes between the two phases.

* Move logic from cacher.go to exporter.go.
* Move tests from cacher_test.go to exporter_test.go.
* Rename cache image flag from '-image' to '-cache-image'.
* Remove unused 'In' field of the Exporter struct.
* Remove cacher implementation, tests, and command.
